refactor(core/server): extract Standard Schema interop into ./server/schema.ts
Moves the schema detection / adaptation / validation layer out of
`server.ts`:

- `isStandard`, `isZodSchema`
- `zodToStandard` (zod → Standard Schema adapter)
- `formatIssues`, `Issue` type (normalised validation-issue shape)
- `validatePart` (params/query/headers/body validator used by the
  handler pipeline)
- `toStandardSchema`, `InferInput`, `InferOutput` (public API)

`server.ts` re-exports the public surface — `formatIssues`,
`toStandardSchema`, `InferInput`, `InferOutput`, `Issue` — so existing
imports keep working. Internal helpers (`isStandard`, `isZodSchema`,
`zodToStandard`) are now module-private to `schema.ts` and no longer
reachable from the top-level entry, which matches their intent.

Also drops the now-unused `requestId` and `~standard`-related imports
that moved along with the helpers.

`wc -l`: server.ts 1833 → 1641; new schema.ts is 205 lines. Tests,
lint, and type-check remain green.
