+++ packages/core/src/server.ts
@@ -18,7 +18,6 @@ import { Scalar } from "@scalar/hono-api-reference";
 import type { StandardSchemaV1 } from "@standard-schema/spec";
 import type { Context, MiddlewareHandler } from "hono";
 import { Hono } from "hono";
-import { requestId } from "hono/request-id";
 import { routePath } from "hono/route";
 import { stream } from "hono/streaming";
 // Type-only Zod imports (optional peer dep)
@@ -51,6 +50,14 @@ import {
 	setCurrentFilePath,
 	setRequestContext,
 } from "./server/context.ts";
+import {
+	formatIssues,
+	type InferInput,
+	type InferOutput,
+	type Issue,
+	toStandardSchema,
+	validatePart,
+} from "./server/schema.ts";
 import { createLogger } from "./utils/logger.ts";
 
 // Re-export the context surface so long-time consumers that imported
@@ -66,6 +73,9 @@ export {
 	setRequestContext,
 };
 export type { LayoutComponent };
+// Re-export schema interop so consumers keep their existing imports.
+export { formatIssues, toStandardSchema };
+export type { InferInput, InferOutput, Issue };
 /**
  * ──────────────────────────────────────────────────────────────────────────────
  * Error + tiny helpers
@@ -568,211 +578,9 @@ export const jsxStreamHono = (
 	});
 };
 
-/**
- * ──────────────────────────────────────────────────────────────────────────────
- * Pure StandardSchemaV1 interop + strong inference
- * ──────────────────────────────────────────────────────────────────────────────
- */
-
-/**
- * Infers the input type from a StandardSchemaV1 schema.
- *
- * @template S - The schema type
- * @returns The inferred input type
- */
-export type InferInput<S> = S extends StandardSchemaV1<unknown, unknown>
-	? StandardSchemaV1.InferInput<S>
-	: S extends ZodTypeAny
-		? z.input<S>
-		: unknown;
-
-/**
- * Infers the output type from a StandardSchemaV1 schema.
- *
- * @template S - The schema type
- * @returns The inferred output type
- */
-export type InferOutput<S> = S extends StandardSchemaV1<unknown, unknown>
-	? StandardSchemaV1.InferOutput<S>
-	: S extends ZodTypeAny
-		? z.output<S>
-		: unknown;
-
-function isStandard(s: unknown): s is AnySchema {
-	return (
-		typeof s === "object" &&
-		s !== null &&
-		typeof (s as { [k: string]: unknown })["~standard"] === "object" &&
-		((s as { [k: string]: unknown })["~standard"] as { version?: unknown })
-			?.version === 1
-	);
-}
-
-// Zod runtime detection (optional)
-function isZodSchema(s: unknown): s is ZodTypeAny {
-	return (
-		!!s &&
-		typeof s === "object" &&
-		typeof (s as { safeParse?: unknown }).safeParse === "function" &&
-		typeof (s as { parse?: unknown }).parse === "function" &&
-		"_def" in (s as object)
-	);
-}
-
-// Adapt a Zod schema to StandardSchemaV1 to reuse common validation
-function zodToStandard<S extends ZodTypeAny>(
-	schema: S,
-	vendor = "hono-file-router",
-): StandardSchemaV1<z.input<S>, z.output<S>> {
-	return {
-		"~standard": {
-			version: 1,
-			vendor,
-			validate: (value: unknown) => {
-				const r = (schema as ZodTypeAny).safeParse(value) as {
-					success: boolean;
-					data?: unknown;
-					error?: { issues?: unknown[] };
-				};
-				if (r.success) return { value: r.data } as const;
-				return { issues: (r.error?.issues ?? []) as unknown[] } as const;
-			},
-		},
-	} as unknown as StandardSchemaV1<z.input<S>, z.output<S>>;
-}
-
-// DX: stable issue shape
-/**
- * Normalised validation issue shape used in error responses.
- */
-export type Issue = {
-	message: string;
-	path: (string | number)[];
-	expected?: string;
-	received?: unknown;
-	type?: string;
-	kind?: string;
-};
-
-function formatIssues(raw: ReadonlyArray<unknown> = []): Issue[] {
-	type LooseIssue = {
-		message?: unknown;
-		path?: unknown;
-		expected?: unknown;
-		received?: unknown;
-		type?: unknown;
-		kind?: unknown;
-	};
-	return raw.map((i) => {
-		const ii = i as LooseIssue;
-		const path = Array.isArray(ii.path)
-			? (ii.path as unknown[]).map((p) => {
-					const pk = (p as { key?: unknown }).key;
-					return typeof pk !== "undefined"
-						? (pk as string | number)
-						: (p as string | number);
-				})
-			: [];
-		return {
-			message: String(ii.message ?? ""),
-			path,
-			expected: typeof ii.expected === "string" ? ii.expected : undefined,
-			received: ii.received,
-			type: typeof ii.type === "string" ? ii.type : undefined,
-			kind: typeof ii.kind === "string" ? ii.kind : undefined,
-		};
-	});
-}
-
-// Per-part validation (params/query/headers/body) with location
-async function validatePart<S extends AnySchema>(
-	where: "params" | "query" | "headers" | "body",
-	schema: S | undefined,
-	value: unknown,
-): Promise<{
-	value: unknown;
-	issues: null | { location: typeof where; issues: Issue[] };
-}> {
-	if (!schema) return { value, issues: null };
-	// Auto-adapt Zod to Standard when provided
-	const effective = isStandard(schema)
-		? schema
-		: isZodSchema(schema)
-			? (zodToStandard(
-					schema as ZodTypeAny,
-					`hono-file-router:${where}`,
-				) as unknown as AnySchema)
-			: schema;
-
-	if (!isStandard(effective)) {
-		return {
-			value: null,
-			issues: {
-				location: where,
-				issues: [
-					{
-						message:
-							"Schema must implement StandardSchemaV1 or be a Zod schema",
-						path: [],
-					},
-				],
-			},
-		};
-	}
-
-	const std = (effective as { [k: string]: unknown })["~standard"] as {
-		validate: (v: unknown) => unknown | Promise<unknown>;
-	};
-	let r = std.validate(value);
-	if (r instanceof Promise) r = await r;
-
-	const fail = r as StandardSchemaV1.FailureResult;
-	if ("issues" in (fail as object)) {
-		return {
-			value: null,
-			issues: { location: where, issues: formatIssues(fail.issues) },
-		};
-	}
-	return {
-		value: (r as StandardSchemaV1.SuccessResult<unknown>).value,
-		issues: null,
-	};
-}
-
-/**
- * Wraps a Valibot schema with the Standard Schema interface used by hono and
- * Scalar so validations and documentation share the same contract.
- *
- * @param schema - Valibot schema to adapt
- * @param vendor - Namespace used to tag the generated schema
- * @returns The schema decorated with StandardSchemaV1 metadata
- */
-export function toStandardSchema<S extends AnySchema>(
-	schema: S,
-	vendor = "hono-file-router",
-): StandardSchemaV1<InferInput<S>, InferOutput<S>> {
-	// Already a Standard schema: just ensure vendor tag
-	if (isStandard(schema)) {
-		const std = (schema as { [k: string]: unknown })["~standard"] as {
-			vendor?: string;
-			[k: string]: unknown;
-		};
-		return {
-			"~standard": {
-				...std,
-				vendor: std.vendor ?? vendor,
-			},
-		} as StandardSchemaV1<InferInput<S>, InferOutput<S>>;
-	}
-	// Zod -> Standard adapter
-	if (isZodSchema(schema)) {
-		return zodToStandard(
-			schema as ZodTypeAny,
-			vendor,
-		) as unknown as StandardSchemaV1<InferInput<S>, InferOutput<S>>;
-	}
-	throw new Error("Schema must implement StandardSchemaV1 or be a Zod schema");
-}
+// Standard Schema interop (zod/valibot detection, adapter, validation,
+// issue formatting) now lives in `./server/schema.ts`. The helpers are
+// imported at the top of this file and re-exported for consumers.
 
 /**
  * ──────────────────────────────────────────────────────────────────────────────
